# docs/contracts/matrix-api.openapi.yaml
openapi: 3.0.3
info:
  title: LeTo Matrix API
  version: 1.0.0
  description: |
    Matrix API for lesson-based exam matrices. totalScore is set directly by the creator (a positive
    integer); each detail row carries a percentage of it, and the sum of percentages must equal 100
    to Submit/Confirm. totalQuestions is derived from details and response-only; delegated matrix
    scope is resolved from the task.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /matrix-tasks:
    post:
      summary: Assign a Matrix task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMatrixTaskRequest' }
      responses:
        '201':
          description: Created Matrix task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatrixTaskResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      summary: List Matrix tasks
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [ASSIGNED, SUBMITTED, COMPLETED] } }
        - { name: dueBefore, in: query, schema: { type: string, format: date-time } }
        - name: keyword
          in: query
          description: Matches the task name or work description, or its id (a plain number).
          schema: { type: string }
        - { name: academicContextId, in: query, schema: { type: integer, format: int64 } }
        - { name: academicYearId, in: query, schema: { type: integer, format: int64 }, description: Lọc độc lập theo năm học của ngữ cảnh }
        - { name: subjectId, in: query, schema: { type: integer, format: int64 } }
        - { name: gradeLevelId, in: query, schema: { type: integer, format: int64 } }
        - { name: semesterId, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Paged Matrix tasks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatrixTaskPage' }
  /my/matrix-tasks:
    get:
      summary: List Matrix tasks assigned to the current Team Lead
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [ASSIGNED, SUBMITTED, COMPLETED] } }
        - { name: dueBefore, in: query, schema: { type: string, format: date-time } }
        - name: keyword
          in: query
          description: Matches the task name or work description, or its id (a plain number).
          schema: { type: string }
        - { name: academicContextId, in: query, schema: { type: integer, format: int64 } }
        - { name: academicYearId, in: query, schema: { type: integer, format: int64 }, description: Lọc độc lập theo năm học của ngữ cảnh }
        - { name: subjectId, in: query, schema: { type: integer, format: int64 } }
        - { name: gradeLevelId, in: query, schema: { type: integer, format: int64 } }
        - { name: semesterId, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Paged assigned tasks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatrixTaskPage' }
  /matrix-tasks/{taskId}:
    get:
      summary: View a Matrix task
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Matrix task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatrixTaskResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete an assigned Matrix task the Team Lead has not started
      description: >
        PHT only (same branch; the Principal any branch). A task counts as started once any matrix exists for it
        (the Team Lead saved a draft or submitted); it can then no longer be deleted.
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 'TaskStarted — "Nhiệm vụ đã được thực hiện, không thể xóa."'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    ValidationError:
      description: Request validation failed
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  schemas:
...
    MatrixPerson:
      type: object
      description: A user shown next to a matrix or task.
      properties:
        userId: { type: integer, format: int64 }
        fullName: { type: string }
        roleLabel:
          type: string
          nullable: true
          description: Hiệu trưởng, Phó Hiệu trưởng or Tổ trưởng; null when the user has no matrix role.
...
    CreateMatrixTaskRequest:
      type: object
      required: [assignedToUserId, academicContextId, name]
      properties:
        assignedToUserId: { type: integer, format: int64 }
        academicContextId: { type: integer, format: int64 }
        semesterId: { type: integer, format: int64, nullable: true }
        dueAt: { type: string, format: date-time, nullable: true }
        name: { type: string, maxLength: 255, description: 'Task name, required — separate from the free-text work description below.' }
        description: { type: string, nullable: true }
    MatrixTaskResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        assignedToUserId: { type: integer, format: int64 }
        academicContextId: { type: integer, format: int64 }
        semesterId: { type: integer, format: int64, nullable: true }
        dueAt: { type: string, format: date-time, nullable: true }
        status: { type: string, example: ASSIGNED }
        taskType: { type: string, example: MATRIX }
        name: { type: string }
        description: { type: string, nullable: true }
        matrixId: { type: integer, format: int64, nullable: true }
        createdBy: { allOf: [{ $ref: '#/components/schemas/MatrixPerson' }], nullable: true, description: The user who assigned the task }
    MatrixTaskPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MatrixTaskResponse' } }
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
...